feat(saving): Wiki with full db can be saved correctly
New save mechanism waits for the sync adaptor to complete their tasks
Removed export database as json functionality (at least on the saver mechanism)
Now it's possible to export the wiki empty
Previously, when you try to download the wiki the sync adaptor starts to fetch all the sikinny tiddlers, but the save mechanism does not allow this task to complete. As a result you end with wikis just partially complete. This does not happen anymore

Fix NoteSelf/NoteSelf.github.io#9